% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constid.R
\name{constid}
\alias{constid}
\title{Get constituency identifier}
\usage{
constid(sourcevar, origin, destination, warn = TRUE)
}
\arguments{
\item{sourcevar}{Vector which contains the codes to be converted}

\item{origin}{Coding scheme to convert from}

\item{destination}{Coding scheme to convert to}

\item{warn}{Flags IDs which return zero or multiple matches}
}
\description{
This function converts between four different sets of constituency
identifiers.
}
\note{
Supports the following coding schemes: Press Association
    reference number ('pa_id'), Hansard ID ('hansard_id'),
    ParlParseID ('parlparse_id'), Office of National Statistics ID
    ('ons_id'). Will try and handle constituency names ('name').
}
\examples{
codes.of.origin <- pano::pano_data$pa_id # Vector of values to be converted
constid(codes.of.origin, "pa_id", "ons_id")
}
\seealso{
\url{election.pressassociation.com/Constituencies/general.php},
    \url{https://en.wikipedia.org/wiki/ONS_coding_system},
    \url{https://github.com/mysociety/parlparse}
}

