% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pano_data.R
\docType{data}
\name{pano_data}
\alias{pano_data}
\title{Constituency Identifier Data Frame}
\format{A data frame with 650 rows and 5 columns}
\usage{
pano_data
}
\description{
A data frame with 650 observations on the following variables.
Used internally by the \code{constid()} function.
}
\details{
\itemize{
  \item hansard_id: Historic Hansard ID
  \item parlparse_id: ID used by the ParlParse project
  \item name: The human-readable name of the constituency
  \item ons_id: ID used by the Office of National Statistics
  \item pa_id: Press Association reference number
}
}
\keyword{datasets}

